Added a new implementation for a debug log component.
Added  methods and macros to use in builds not meant for distribution.
The macros work both in debug and in non debug mode.
Include the files:

or

To use this feature. tools should be included in projects built after toosl, while
osl should be included in project built after sal (limited functionality, though).

The value DebugLoggerFileAgeInHours in Bootstrap section,
in setuprc/setup.ini can be used to adapt the log file age
in hours, for example:

DebugLoggerFileAgeInHours=24

means the file aged more than 24 hour from the LO startup time,
will be cancelled.

A list of subprojects involved:

[sal, tools] Added new methods to have the debug trace sent to a file.
[offapi] Add the new interface and singleton implementation of the new debug log component
[comphelper] Add the new debug logging component
[tools] Add helper methods and macros for debug logs in tools
[desktop] Added main start and flush for debug log event recorder and some logs
[sal] Added OS independent functions to serve new debug logger methods availble in tools

Added a new interface and singleton implementation for a debug log component: squash this commit