i125194 Added two missing win32 API error definitions.
ERROR_FILE_CHECKED_OUT
returned when trying to access a local filesystem file that is locked by another user.
Works only if the local file is managed by the 'Web Client Service' service.

ERROR_INVALID_NAME
returned when trying to access a local filesystem file with the wrong chars in the filename.
Works only if the local file is managed by the 'Web Client Service' service.